Move commands to thirdparty folder
